+++ app/src/main/java/cc/kafuu/bilidownload/common/utils/TimeUtils.kt
@@ -1,7 +1,7 @@
 package cc.kafuu.bilidownload.common.utils
 
 object TimeUtils {
-    fun formatSecondTime(seconds: Int): String {
+    fun formatSecondTime(seconds: Long): String {
         val hours = seconds / 3600
         val minutes = (seconds % 3600) / 60
         val secs = seconds % 60
